packages/git-storage-sdk-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pierre-storage"
version = "0.12.1"
description = "Pierre Git Storage SDK for Python"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Pierre", email = "support@pierre.io"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.27.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.5.0; python_version < '3.10'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.urls]
Homepage = "https://pierre.io"
Documentation = "https://docs.pierre.io"
Repository = "https://github.com/pierrecomputer/pierre"
Issues = "https://github.com/pierrecomputer/pierre/issues"

[tool.setuptools.packages.find]
include = ["pierre_storage*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=pierre_storage --cov-report=term-missing"

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]
constraint-dependencies = [
    "urllib3>=2.6.3",
]
